Reverted branch 1.3.1 to r1209 as that was the released version
git-svn-id: https://duplicati.googlecode.com/svn/branches/1.3.1@1216 59da171f-624f-0410-aa54-27559c288bec

+++ Duplicati/Library/Backend/GoogleDocs/GoogleDocs.cs
@@ -141,7 +141,7 @@ namespace Duplicati.Library.Backend
                         break;
                     }
 
-                results.Add(file.Title, new TaggedFileEntry(file.Title, (long)file.QuotaBytesUsed, file.LastViewed, file.Updated, file.ResourceId, file.DocumentEntry.Content.Src.ToString(), updateUrl, file.ETag));
+                results.Add(file.Title, new TaggedFileEntry(file.Title, file.QuotaBytesUsed, file.LastViewed, file.Updated, file.ResourceId, file.DocumentEntry.Content.Src.ToString(), updateUrl, file.ETag));
             }
             return results;
         }
